# docs/archive/openapi.root.skeleton.yaml
openapi: 3.0.3
info:
  title: Shapemint
  version: 0.1.0
  description: Influence-and-incentive ledger measuring third-party impact on AI model behaviour.
  x-product-slug: ai-decentralized-shapemint
  x-source: ai-in-decentralized+ai/decentralizedai-boston-181007051624/
servers:
- url: https://api.example.com
...
paths:
...
  /v1/governance/minting-pause:
    post:
      operationId: pauseMinting
      tags:
      - Governance
      security:
      - BearerAuth: []
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeMintingPause'
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MintingPause'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    ResponseMeta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    MintingPause:
      type: object
      required:
      - id
      - modelId
      - active
      properties:
        id:
          type: string
        modelId:
          type: string
        active:
          type: boolean
        reason:
          type: string
...
    DataEnvelopeMintingPause:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/MintingPause'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...