# platform/api-server/src/domains/governance/handlers/minting-pause.handlers.ts
/**
 * Handlers for MintingPause
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: governance
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { GovernanceDomainModule } from "../dependencies/governance-ddd.dependencies.js";
import type { MintingPauseCreate } from "@shapemint/core/governance/types";
import type { ListMintingPausesInput, PauseMintingInput } from "@shapemint/services/governance";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/governance/minting-pause
 * Operation: pauseMinting
 */
export async function pauseMinting(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: Record<string, string | string[] | undefined>; Body: MintingPauseCreate }>,
  reply: FastifyReply,
  deps: GovernanceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: PauseMintingInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.mintingPauses.get.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/governance/minting-pauses
 * Operation: listMintingPauses
 */
export async function listMintingPauses(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], modelId?: string | string[], active?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: GovernanceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListMintingPausesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.mintingPauses.list.execute(input);
  return reply.code(200).send(result);
}